\name{topcolour-ig-vancouver-2014}
\alias{topcolour.ig.vancouver.2014}
\docType{data}
\title{
Instagram 2014 Vancouver dataset top colour
}
\description{
Top Colour of all Vancouver, BC Instagram 2014 thumbnails
}
\usage{data("topcolour.ig.vancouver.2014")}
\format{
  A data frame with 245736 observations on the following 2 variables: top colour and count. The top colour is the top colour from Vancouver, BC, Canada Instagram images taken in 2014.
\describe{
\item{colour}{colour, in RGB Hex}
\item{count}{count, the number of times that top colour occurs}
}
}
\keyword{datasets}
